+++ Logic/Environment.py
@@ -243,8 +243,17 @@ class Environment:
     # Close system objects
     def close(self):
         # This will try to safely shut down any objects that are capable of running threads.
+        self.closeRobot()
+        self.closeVideo()
+
+    def closeRobot(self):
+        self.__robot.disconnect()
         self.__robot.setExiting(True)
+
+    def closeVideo(self):
         self.__vision.setExiting(True)
         self.__vStream.endThread()
-
+        self.__vStream.frame = None
+        self.__vStream.cap = None
+        self.__vStream.setCamera = None